dove mancava cambiati i puntatori da tipo *nome a tipo* nome; aggiunti i flag C99 come da richiesta

+++ src/utils.c
@@ -93,7 +93,7 @@ bool konamiCode() {
     const char konamiSequence[] = {'w','s','s','a','d','a','d','b','a',' '};
     char code;
     
-    for(int i = 0; i < sizeof(konamiSequence); i++) {
+    for(size_t i = 0; i < sizeof(konamiSequence); i++) {
         printf("Seleziona una delle opzioni [1-3]: ");
         code = readOption("123wsadb ");
